+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `Press` into `JustPress`.
+
 ## [0.2.0] - 2024-11-03
 
 ### Added